Fix maps not loading when container runs
Move to specifying the maprotation.cfg via +exec in the entrypoint as
having the exec in  q3server_config.cfg didn't actually ever load the maps.

Add Makefile code to randomize the rcon_password and subsitute it
into the q3server_config.cfg.tmpl, then generate a q3server_config.cfg
in output and build that into the Docker container.